# perf: precalculate fragment and scheme state