--- a/src/character_sets.cpp
+++ b/src/character_sets.cpp
@@ -436,7 +436,7 @@ namespace ada::character_sets {
       0x01 | 0x02 | 0x04 | 0x08 | 0x10 | 0x20 | 0x40 | 0x80
   };
 
-  bool bit_at(const uint8_t a[], const uint8_t i) {
+  ada_really_inline bool bit_at(const uint8_t a[], const uint8_t i) {
     return !!(a[i >> 3] & (1 << (i & 7)));
   }