--- a/src/checkers.cpp
+++ b/src/checkers.cpp
@@ -1,5 +1,4 @@
 #include "ada.h"
-#include "unicode.cpp"
 
 #include <algorithm>