--- a/src/unicode.cpp
+++ b/src/unicode.cpp
@@ -1,5 +1,3 @@
-#include <unordered_set>
-
 namespace ada::unicode {
 
   // A forbidden host code point is U+0000 NULL, U+0009 TAB, U+000A LF, U+000D CR, U+0020 SPACE, U+0023 (#),