Before:

```
BasicBench_AdaURL           12170 ns        12165 ns        57778 time/byte=15.0553ns
```

After:
```
BasicBench_AdaURL            7513 ns         7513 ns        93508 time/byte=9.29769ns
```